create new file for utils used by start_tasks.py
